pyproject.toml
[project]
name = "music-recommendation-system"
version = "0.1.0"
description = "Production-ready music recommendation system with classical, ML, and deep-learning pipelines."
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "duckdb>=1.4.4",
    "hydra-core>=1.3.2",
    "jupyterlab>=4.5.3",
    "matplotlib>=3.10.8",
    "mlflow>=3.5.1",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.2.0",
    "plotly>=6.3.1",
    "polars[pandas]>=1.37.1",
    "pygments>=2.20.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "tornado>=6.5.5",
    "torch>=2.9.0",
    "typer>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "bandit[toml]>=1.8.6",
    "ipykernel>=7.0.1",
    "pip-audit>=2.9.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
]

[project.scripts]
music-rec = "music_recommendation.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/music_recommendation"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "artifacts", "notebooks"]
skips = ["B101"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
